% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{Chal_func}
\alias{Chal_func}
\title{Challenge Function}
\usage{
Chal_func(x)
}
\arguments{
\item{x}{a nummeric value}
}
\value{
numeric
}
\description{
Challenge Function
}
\examples{
Chal_func(15)         ##dieser Teil wurde in den Markdown "Challenge" eingebunden


}
