% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{followers}
\alias{followers}
\title{followers}
\usage{
followers(x)
}
\arguments{
\item{x}{any integer value}
}
\value{
the same value and the two following
}
\description{
followers
}
\examples{
followers(3)
}
