% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{new_fun}
\alias{new_fun}
\title{Test Function}
\usage{
new_fun(x, y)
}
\arguments{
\item{x}{numeric value}

\item{y}{numeric value}
}
\value{
numeric value
}
\description{
Test Function
}
\examples{

c <- new_fun(3, 5.4)
c

}
