% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pipeme}
\alias{pipeme}
\title{pipeme}
\usage{
pipeme(dataset, column, value)
}
\arguments{
\item{dataset}{a dataframe}

\item{column}{a columncontent as character}
}
\value{
a column vector
}
\description{
pipeme
}
\examples{
asd=data.frame(Values=c(1,2,3,4),else=c(2,3,4,5))
pipeme(asd,"val")

}
