% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{returnname}
\alias{returnname}
\title{returnname}
\usage{
returnname(x)
}
\arguments{
\item{x}{a character string}
}
\value{
a returned character string
}
\description{
returnname
}
\examples{
returnname("Wolfgang")
}
