% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{tobler_fun}
\alias{tobler_fun}
\title{Tobler Function for off-path walking}
\usage{
tobler_fun(x)
}
\arguments{
\item{x}{numeric values; e.g. raster values of digital elevation model}
}
\value{
numeric values
}
\description{
Tobler Function for off-path walking
}
\examples{
(raster package required)
rast <- raster(nrow=10, ncol=10)
rast[] <- runif(ncell(rast))
x <- tobler_fun(rast[])
head(x)
}
